Bind keywords in the runtime compilers and shrink the legacy surface
The browser compilers now bind every keyword to a bare local name, which
loose and strict templates can both invoke, and the evaluators supply
the values. That takes `compile()`, the test harness, and runtime
`template()` off the built-in name tables. The tables and their fill
remain only behind `@ember/template-factory`, for JSON blocks that
addons published before importable opcodes.

`-resolve` and `-disallow-dynamic-resolution` bind like every other
keyword, and the `{{#each-in}}` iterator moves into the helper module
that owns it.
