+++ packages/@ember/template-compiler/lib/-internal/primitives.ts
@@ -1,1 +1,2 @@
 export { RESOLUTION_MODE_TRANSFORMS, STRICT_MODE_TRANSFORMS, INTERNAL_PLUGINS } from '../plugins';
+export { lookupRuntimeKeyword, RUNTIME_KEYWORD_LOCALS } from '../compile-options';
