+++ packages/ember-template-compiler/lib/public-api.ts
@@ -16,5 +16,7 @@ export type { EmberPrecompileOptions } from './types';
 
 export { preprocess as _preprocess, print as _print } from '@glimmer/syntax';
 export { precompile as _precompile } from '@glimmer/compiler';
+export { RUNTIME_KEYWORD_LOCALS } from '@ember/template-compiler/-internal-primitives';
+export { withRuntimeKeywords } from './system/runtime-keywords';
 
 export { _GlimmerSyntax, VERSION };
