+++ tests/node-vitest/tree-shakability.test.js
@@ -34,6 +34,7 @@ it('[dev] source modules with side effects', () => {
       "@ember/array/lib/iterator-extension.js",
       "@ember/debug/index.js",
       "@ember/runloop/index.js",
+      "@ember/template-compiler/lib/compile-options.js",
       "@glimmer/compiler/lib/wire-format-module.js",
       "@glimmer/validator/index.js",
       "@handlebars/parser/lib/parse.js",
@@ -60,6 +61,7 @@ it('[prod] source modules with side effects', () => {
       "@ember/array/index.js",
       "@ember/array/lib/iterator-extension.js",
       "@ember/runloop/index.js",
+      "@ember/template-compiler/lib/compile-options.js",
       "@glimmer/compiler/lib/wire-format-module.js",
       "@glimmer/validator/index.js",
       "@handlebars/parser/lib/parse.js",
